Refactor test classes
