+++ app/src/test/java/co/rahulchowdhury/elly/seed/elephant/SeedData.kt
@@ -1,0 +1,11 @@
+package co.rahulchowdhury.elly.seed.elephant
+
+import co.rahulchowdhury.elly.data.model.local.Elephant
+
+val seedElephant = Elephant(
+    id = "1",
+    name = "Elly",
+    affiliation = "Happy Feet",
+    lastFetchTime = 0,
+    image = "elly-profile.jpg"
+)
