+++ app/src/test/java/co/rahulchowdhury/elly/ui/profile/ElephantProfileViewModelTest.kt
@@ -1,34 +1,18 @@
 package co.rahulchowdhury.elly.ui.profile
 
-import androidx.arch.core.executor.testing.InstantTaskExecutorRule
 import co.rahulchowdhury.elly.LiveDataTestUtil.getValue
-import co.rahulchowdhury.elly.MainCoroutineRule
-import co.rahulchowdhury.elly.data.model.local.Elephant
 import co.rahulchowdhury.elly.data.repo.FakeElephantRepository
 import co.rahulchowdhury.elly.exception.base.InvalidArgumentException
+import co.rahulchowdhury.elly.seed.elephant.seedElephant
+import co.rahulchowdhury.elly.ui.base.ViewModelTest
 import com.google.common.truth.Truth.assertThat
-import kotlinx.coroutines.ExperimentalCoroutinesApi
 import org.junit.Before
-import org.junit.Rule
 import org.junit.Test
 
-class ElephantProfileViewModelTest {
+class ElephantProfileViewModelTest : ViewModelTest() {
     private lateinit var elephantProfileViewModel: ElephantProfileViewModel
     private lateinit var fakeElephantRepository: FakeElephantRepository
-    private val elephant = Elephant(
-        id = "1",
-        name = "Elly",
-        affiliation = "Happy Feet",
-        lastFetchTime = 0,
-        image = "elly-profile.jpg"
-    )
-
-    @ExperimentalCoroutinesApi
-    @get:Rule
-    var mainCoroutineRule = MainCoroutineRule()
-
-    @get:Rule
-    var instantExecutorRule = InstantTaskExecutorRule()
+    private val elephant = seedElephant
 
     @Before
     fun setUp() {
